backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4
pytest-flask==1.3.0
gunicorn==21.2.0
requests==2.31.0
lxml==5.1.0
beautifulsoup4==4.12.3
openpyxl==3.1.2
